+++ R/globals.R
@@ -1,7 +1,8 @@
 # Suppress R CMD check notes about undefined global variables
 utils::globalVariables(c(
-  "Individual_Variance", 
-  "Cumulative_Variance", 
-  "Exclude", 
-  "Timepoint"
+  "Individual_Variance",
+  "Cumulative_Variance",
+  "Exclude",
+  "Timepoint",
+  ".unit_id"   # internal well-identity column built by .nova_unit_id()
 ))